# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "York academy",
        "description": "York academy",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/"
        },
        {
            "url": "/api/"
        }
    ],
    "paths": {
...
        "/plan_register": {
            "get": {
                "tags": [
                    "Plan Registers"
                ],
                "summary": "Retrieve all plan registers",
                "description": "Get a list of all plan registers",
                "operationId": "getAllPlanRegisters",
                "responses": {
                    "200": {
                        "description": "List of plan registers retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PlanRegister"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Plan Registers"
                ],
                "summary": "Store a new registration plan",
                "description": " new registration plan with provided details",
                "operationId": "storePlanRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "training_plan_id": {
                                        "description": "ID of the training plan",
                                        "type": "integer",
                                        "example": null
                                    },
                                    "full_name": {
                                        "description": "Full name of the user",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "description": "Phone number of the user",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "email": {
                                        "description": "Email address of the user",
                                        "type": "string",
                                        "example": "omaralsayyad20032003@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration plan stored successfully"
                    }
                }
            }
        },
        "/plan_register/{id}": {
            "get": {
                "tags": [
                    "Plan Registers"
                ],
                "summary": "Retrieve a specific plan register",
                "description": "Get details of a specific plan register by ID",
                "operationId": "getPlanRegisterById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the plan register to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan register retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PlanRegister"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plan register not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plan register not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PlanRegister": {
                "description": "Plan register model",
                "properties": {
                    "id": {
                        "description": "ID of the registration",
                        "type": "integer"
                    },
                    "training_plan_id": {
                        "description": "ID of the training plan",
                        "type": "integer"
                    },
                    "full_name": {
                        "description": "Full name of the user",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Phone number of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address of the user",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
...
    },
...
}